requirements.txt
# Copyright 2021 Wirepas Ltd 2021

wirepas-mqtt-library==1.2.5
cbor2==5.6.4
PyYAML==6.0.1
pycryptodome==3.20.0
mypy==1.10.1